% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colour_words}
\alias{colour_words}
\title{Hex colours which are directly readable as words}
\format{
An object of class \code{character} of length 24.
}
\usage{
colour_words
}
\description{
Hex colours which are directly readable as words
}
\keyword{datasets}
