% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colour_words_with_substitutions}
\alias{colour_words_with_substitutions}
\title{Hex colours which are readable as words with some digit-to-letter substitutions}
\format{
An object of class \code{character} of length 1075.
}
\usage{
colour_words_with_substitutions
}
\description{
Hex colours which are readable as words with some digit-to-letter substitutions
}
\keyword{datasets}
