% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_best_colour_match}
\alias{find_best_colour_match}
\title{Find best matches for a set of reference colours}
\usage{
find_best_colour_match(reference_colours, candidate_colours)
}
\arguments{
\item{reference_colours}{original colours}

\item{candidate_colours}{large list of possible candidate colours. In this
case these are colours possibly generated from words}
}
\value{
character vector of colours in \code{candidate_colours} which best
        match the colours in \code{reference_colours}
}
\description{
Find best matches for a set of reference colours
}
