% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonsense_words}
\alias{nonsense_words}
\title{Hex colours which are readable as nonsense words with some digit-to-letter substitutions}
\format{
An object of class \code{character} of length 28651.
}
\usage{
nonsense_words
}
\description{
Hex colours which are readable as nonsense words with some digit-to-letter substitutions
}
\keyword{datasets}
