% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_palette}
\alias{show_palette}
\title{show a single palette with a title}
\usage{
show_palette(
  colours,
  title = NULL,
  labels = NULL,
  show_labels = TRUE,
  borders = NULL,
  cex_label = 1,
  ncol = NULL
)
}
\arguments{
\item{colours}{character vector of hex colours}

\item{title}{title}

\item{labels}{colour labels. if null then use colour names as-is, otherwise
use these labels given as a character vector}

\item{show_labels}{default: TRUE}

\item{borders}{set borders}

\item{cex_label}{size for colour label}

\item{ncol}{number of colums. id NULL (the default), then make the result
as square as possible}
}
\description{
show a single palette with a title
}
