% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-words-colours.R
\name{words_to_colours}
\alias{words_to_colours}
\alias{colours_to_words}
\title{Cnvert words to colours using the given number to letter subsitituions}
\usage{
words_to_colours(
  words,
  substitutions = c(o = 0, l = 1, z = 2, s = 5, t = 7, g = 9)
)

colours_to_words(
  colours,
  substitutions = c(o = 0, l = 1, z = 2, s = 5, t = 7, g = 9)
)
}
\arguments{
\item{words}{character vector of words which can be represented as colours}

\item{substitutions}{number-to-letter substitutions substituions}

\item{colours}{character vector of colours}
}
\description{
Cnvert words to colours using the given number to letter subsitituions
}
